+++ components/mdz/internal/rest/organization_test.go
@@ -4,93 +4,59 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/LerianStudio/midaz/components/mdz/internal/model"
+	"github.com/LerianStudio/midaz/common/mmodel"
 	"github.com/LerianStudio/midaz/components/mdz/pkg/environment"
 	"github.com/LerianStudio/midaz/components/mdz/pkg/factory"
+	"github.com/LerianStudio/midaz/components/mdz/pkg/mockutil"
 	"github.com/LerianStudio/midaz/components/mdz/pkg/ptr"
 	"github.com/jarcoal/httpmock"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestCreateOrganization(t *testing.T) {
-
-	mockResp := `{
-		"id": "1a259e90-8f28-491d-8c09-c047293b1a0f",
-		"legalName": "Corwin LLC",
-		"doingBusinessAs": "The ledger.io",
-		"legalDocument": "48784548000104",
-		"address": {
-			"line1": "Avenida Paulista, 1234",
-			"line2": "CJ 203",
-			"zipCode": "04696040",
-			"city": "Ozellaport",
-			"state": "MU",
-			"country": "LV"
-		},
-		"status": {
-			"code": "ACTIVE",
-			"description": "Teste Ledger"
-		},
-		"createdAt": "2024-10-22T23:36:35.979331542Z",
-		"updatedAt": "2024-10-22T23:36:35.979334336Z",
-		"metadata": {
-			"bitcoinn": "3fozQHVceartTg14kwF4PkfgUU4JhsUX",
-			"boolean": true,
-			"chave": "metadata_chave",
-			"double": 10.5,
-			"int": 1
-		}
-	}`
-
-	input := model.Organization{
-		LegalName:       "Corwin LLC",
-		DoingBusinessAs: "The ledger.io",
-		LegalDocument:   "48784548000104",
-		Status: model.Status{
-			Code:        ptr.StringPtr("ACTIVE"),
-			Description: "Teste Ledger",
-		},
-		Address: model.Address{
-			Line1:   ptr.StringPtr("Avenida Paulista, 1234"),
-			Line2:   ptr.StringPtr("CJ 203"),
-			ZipCode: ptr.StringPtr("04696040"),
-			City:    ptr.StringPtr("Ozellaport"),
-			State:   ptr.StringPtr("MU"),
-			Country: "LV",
-		},
-		Metadata: &model.Metadata{
-			Chave:   ptr.StringPtr("metadata_chave"),
-			Bitcoin: ptr.StringPtr("3fozQHVceartTg14kwF4PkfgUU4JhsUX"),
-			Boolean: ptr.BoolPtr(true),
-			Double:  ptr.Float64Ptr(10.5),
-			Int:     ptr.IntPtr(1),
-		},
+	metadata := map[string]any{
+		"bitcoin": "3fozQHVceartTg14kwF4PkfgUU4JhsUX",
+		"chave":   "metadata_chave",
+		"boolean": true,
 	}
 
-	expectedResult := &model.OrganizationCreate{
-		ID:              "1a259e90-8f28-491d-8c09-c047293b1a0f",
+	inp := mmodel.CreateOrganizationInput{
 		LegalName:       "Corwin LLC",
-		DoingBusinessAs: "The ledger.io",
+		DoingBusinessAs: ptr.StringPtr("The ledger.io"),
 		LegalDocument:   "48784548000104",
-		Address: model.Address{
-			Line1:   ptr.StringPtr("Avenida Paulista, 1234"),
+		Status: mmodel.Status{
+			Code:        "ACTIVE",
+			Description: ptr.StringPtr("Teste Ledger"),
+		},
+		Address: mmodel.Address{
+			Line1:   "Avenida Paulista, 1234",
 			Line2:   ptr.StringPtr("CJ 203"),
-			ZipCode: ptr.StringPtr("04696040"),
-			City:    ptr.StringPtr("Ozellaport"),
-			State:   ptr.StringPtr("MU"),
+			ZipCode: "04696040",
+			City:    "Ozellaport",
+			State:   "MU",
 			Country: "LV",
 		},
-		Status: model.Status{
-			Code:        ptr.StringPtr("ACTIVE"),
-			Description: "Teste Ledger",
+		Metadata: metadata,
+	}
+
+	expectedResult := mmodel.Organization{
+		ID:              "1a259e90-8f28-491d-8c09-c047293b1a0f",
+		LegalName:       "Corwin LLC",
+		DoingBusinessAs: ptr.StringPtr("The ledger.io"),
+		LegalDocument:   "48784548000104",
+		Address: mmodel.Address{
+			Line1:   "Avenida Paulista, 1234",
+			Line2:   ptr.StringPtr("CJ 203"),
+			ZipCode: "04696040",
+			City:    "Ozellaport",
+			State:   "MU",
+			Country: "LV",
 		},
-		Metadata: model.Metadata{
-			Chave:   ptr.StringPtr("metadata_chave"),
-			Bitcoin: ptr.StringPtr("3fozQHVceartTg14kwF4PkfgUU4JhsUX"),
-			Boolean: ptr.BoolPtr(true),
-			Double:  ptr.Float64Ptr(10.5),
-			Int:     ptr.IntPtr(1),
+		Status: mmodel.Status{
+			Code:        "ACTIVE",
+			Description: ptr.StringPtr("Teste Ledger"),
 		},
+		Metadata: metadata,
 	}
 
 	client := &http.Client{}
@@ -100,7 +66,7 @@ func TestCreateOrganization(t *testing.T) {
 	URIAPILedger := "http://127.0.0.1:3000"
 
 	httpmock.RegisterResponder(http.MethodPost, URIAPILedger+"/v1/organizations",
-		httpmock.NewStringResponder(http.StatusCreated, mockResp))
+		mockutil.MockResponseFromFile(http.StatusCreated, "./.fixtures/organization_response_create.json"))
 
 	factory := &factory.Factory{
 		HTTPClient: client,
@@ -109,62 +75,42 @@ func TestCreateOrganization(t *testing.T) {
 		},
 	}
 
-	orgService := NewOrganization(factory)
+	org := NewOrganization(factory)
 
-	result, err := orgService.Create(input)
+	result, err := org.Create(inp)
 
 	assert.NoError(t, err)
+
 	assert.NotNil(t, result)
 	assert.Equal(t, expectedResult.ID, result.ID)
 	assert.Equal(t, expectedResult.LegalName, result.LegalName)
 	assert.Equal(t, expectedResult.DoingBusinessAs, result.DoingBusinessAs)
 	assert.Equal(t, expectedResult.LegalDocument, result.LegalDocument)
 	assert.Equal(t, expectedResult.Address.Line1, result.Address.Line1)
+	assert.Equal(t, expectedResult.Metadata, result.Metadata)
 
 	info := httpmock.GetCallCountInfo()
 	assert.Equal(t, 1, info["POST http://127.0.0.1:3000/v1/organizations"])
 }
 
 func TestOrganizationGetByID(t *testing.T) {
-
-	mockResp := `{
-	    "id": "0192c559-62f4-738b-9be5-262b71f6375a",
-	    "parentOrganizationId": null,
-	    "legalName": "Koelpin - Marquardt",
-	    "doingBusinessAs": "The ledger.io",
-	    "legalDocument": "48784548000104",
-	    "address": {
-	        "line1": "Avenida Paulista, 1234",
-	        "line2": null,
-	        "zipCode": "",
-	        "city": "",
-	        "state": "",
-	        "country": "KG"
-	    },
-	    "status": {
-	        "code": "ACTIVE",
-	        "description": "Teste Ledger"
-	    },
-	    "createdAt": "2024-10-25T20:23:42.580221Z",
-	    "updatedAt": "2024-10-25T20:23:42.580221Z",
-	    "deletedAt": null,
-	    "metadata": {
-	        "bitcoinn": "1R4DvodZi68SxKbvNeQGCkaPj25Ryumy",
-	        "chave": "max"
-	    }
-	}`
-
 	organizationID := "0192c559-62f4-738b-9be5-262b71f6375a"
 	URIAPILedger := "http://127.0.0.1:3000"
 
-	expectedResult := &model.OrganizationItem{
+	metadata := map[string]any{
+		"bitcoin": "1R4DvodZi68SxKbvNeQGCkaPj25Ryumy",
+		"chave":   "max",
+	}
+
+	expectedResult := &mmodel.Organization{
 		ID:              organizationID,
 		LegalName:       "Koelpin - Marquardt",
-		DoingBusinessAs: "The ledger.io",
+		DoingBusinessAs: ptr.StringPtr("The ledger.io"),
 		LegalDocument:   "48784548000104",
-		Address: model.Address{
-			Line1: ptr.StringPtr("Avenida Paulista, 1234"),
+		Address: mmodel.Address{
+			Line1: "Avenida Paulista, 1234",
 		},
+		Metadata: metadata,
 	}
 
 	client := &http.Client{}
@@ -172,7 +118,7 @@ func TestOrganizationGetByID(t *testing.T) {
 	defer httpmock.DeactivateAndReset()
 
 	httpmock.RegisterResponder(http.MethodGet, URIAPILedger+"/v1/organizations/"+organizationID,
-		httpmock.NewStringResponder(http.StatusOK, mockResp))
+		mockutil.MockResponseFromFile(http.StatusOK, "./.fixtures/organization_response_get_by_id.json"))
 
 	factory := &factory.Factory{
 		HTTPClient: client,
@@ -181,9 +127,9 @@ func TestOrganizationGetByID(t *testing.T) {
 		},
 	}
 
-	orgService := NewOrganization(factory)
+	org := NewOrganization(factory)
 
-	result, err := orgService.GetByID("0192c559-62f4-738b-9be5-262b71f6375a")
+	result, err := org.GetByID("0192c559-62f4-738b-9be5-262b71f6375a")
 	assert.NoError(t, err)
 	assert.NotNil(t, result)
 	assert.Equal(t, expectedResult.ID, result.ID)
@@ -191,67 +137,46 @@ func TestOrganizationGetByID(t *testing.T) {
 	assert.Equal(t, expectedResult.DoingBusinessAs, result.DoingBusinessAs)
 	assert.Equal(t, expectedResult.LegalDocument, result.LegalDocument)
 	assert.Equal(t, expectedResult.Address.Line1, result.Address.Line1)
+	assert.Equal(t, expectedResult.Metadata, result.Metadata)
 
 	info := httpmock.GetCallCountInfo()
 	assert.Equal(t, 1, info["GET http://127.0.0.1:3000/v1/organizations/0192c559-62f4-738b-9be5-262b71f6375a"])
 }
 
 func TestOrganizationUpdate(t *testing.T) {
-	mockResp := `{
-		"id": "1a259e90-8f28-491d-8c09-c047293b1a0f",
-		"legalName": "Corwin LLC",
-		"doingBusinessAs": "The ledger.io",
-		"legalDocument": "48784548000104",
-		"address": {
-			"line1": "Avenida Paulista, 1234",
-			"line2": "CJ 203",
-			"zipCode": "04696040",
-			"city": "Ozellaport",
-			"state": "MU",
-			"country": "LV"
-		},
-		"status": {
-			"code": "ACTIVE",
-			"description": "Teste Ledger"
-		},
-		"createdAt": "2024-10-22T23:36:35.979331542Z",
-		"updatedAt": "2024-10-22T23:36:35.979334336Z"
-	}`
-
-	input := model.OrganizationUpdate{
+	inp := mmodel.UpdateOrganizationInput{
 		LegalName:       "Corwin LLC",
-		DoingBusinessAs: "The ledger.io",
-		LegalDocument:   ptr.StringPtr("48784548000104"),
-		Status: &model.StatusUpdate{
-			Code:        ptr.StringPtr("ACTIVE"),
+		DoingBusinessAs: ptr.StringPtr("The ledger.io"),
+		Status: mmodel.Status{
+			Code:        "ACTIVE",
 			Description: ptr.StringPtr("Teste Ledger"),
 		},
-		Address: model.Address{
-			Line1:   ptr.StringPtr("Avenida Paulista, 1234"),
+		Address: mmodel.Address{
+			Line1:   "Avenida Paulista, 1234",
 			Line2:   ptr.StringPtr("CJ 203"),
-			ZipCode: ptr.StringPtr("04696040"),
-			City:    ptr.StringPtr("Ozellaport"),
-			State:   ptr.StringPtr("MU"),
+			ZipCode: "04696040",
+			City:    "Ozellaport",
+			State:   "MU",
 			Country: "LV",
 		},
 	}
 
-	expectedResult := &model.OrganizationItem{
+	expectedResult := &mmodel.Organization{
 		ID:              "1a259e90-8f28-491d-8c09-c047293b1a0f",
 		LegalName:       "Corwin LLC",
-		DoingBusinessAs: "The ledger.io",
+		DoingBusinessAs: ptr.StringPtr("The ledger.io"),
 		LegalDocument:   "48784548000104",
-		Address: model.Address{
-			Line1:   ptr.StringPtr("Avenida Paulista, 1234"),
+		Address: mmodel.Address{
+			Line1:   "Avenida Paulista, 1234",
 			Line2:   ptr.StringPtr("CJ 203"),
-			ZipCode: ptr.StringPtr("04696040"),
-			City:    ptr.StringPtr("Ozellaport"),
-			State:   ptr.StringPtr("MU"),
+			ZipCode: "04696040",
+			City:    "Ozellaport",
+			State:   "MU",
 			Country: "LV",
 		},
-		Status: model.Status{
-			Code:        ptr.StringPtr("ACTIVE"),
-			Description: "Teste Ledger",
+		Status: mmodel.Status{
+			Code:        "ACTIVE",
+			Description: ptr.StringPtr("Teste Ledger"),
 		},
 	}
 
@@ -263,7 +188,7 @@ func TestOrganizationUpdate(t *testing.T) {
 	organizationID := "1a259e90-8f28-491d-8c09-c047293b1a0f"
 
 	httpmock.RegisterResponder(http.MethodPatch, URIAPILedger+"/v1/organizations/"+organizationID,
-		httpmock.NewStringResponder(http.StatusOK, mockResp))
+		mockutil.MockResponseFromFile(http.StatusOK, "./.fixtures/organization_response_update.json"))
 
 	factory := &factory.Factory{
 		HTTPClient: client,
@@ -272,9 +197,9 @@ func TestOrganizationUpdate(t *testing.T) {
 		},
 	}
 
-	orgService := NewOrganization(factory)
+	org := NewOrganization(factory)
 
-	result, err := orgService.Update(organizationID, input)
+	result, err := org.Update(organizationID, inp)
 
 	assert.NoError(t, err)
 	assert.NotNil(t, result)
@@ -306,12 +231,74 @@ func Test_organization_Delete(t *testing.T) {
 		},
 	}
 
-	orgService := NewOrganization(factory)
+	org := NewOrganization(factory)
 
-	err := orgService.Delete(organizationID)
+	err := org.Delete(organizationID)
 
 	assert.NoError(t, err)
 
 	info := httpmock.GetCallCountInfo()
 	assert.Equal(t, 1, info["DELETE http://127.0.0.1:3000/v1/organizations/1a259e90-8f28-491d-8c09-c047293b1a0f"])
+}
+
+func Test_organization_Get(t *testing.T) {
+	organizationID := "0192c559-62f4-738b-9be5-262b71f6375a"
+	URIAPILedger := "http://127.0.0.1:3000"
+
+	limit := 5
+	page := 1
+
+	metadata := map[string]any{
+		"bitcoin": "1R4DvodZi68SxKbvNeQGCkaPj25Ryumy",
+		"chave":   "max",
+	}
+
+	item := mmodel.Organization{
+		ID:              organizationID,
+		LegalName:       "Koelpin - Marquardt",
+		DoingBusinessAs: ptr.StringPtr("The ledger.io"),
+		LegalDocument:   "48784548000104",
+		Address: mmodel.Address{
+			Line1: "Avenida Paulista, 1234",
+		},
+		Metadata: metadata,
+	}
+
+	expectedResult := mmodel.Organizations{
+		Items: []mmodel.Organization{
+			item,
+		},
+		Limit: limit,
+		Page:  page,
+	}
+
+	client := &http.Client{}
+	httpmock.ActivateNonDefault(client)
+	defer httpmock.DeactivateAndReset()
+
+	httpmock.RegisterResponder(http.MethodGet, URIAPILedger+"/v1/organizations?limit=5&page=1",
+		mockutil.MockResponseFromFile(http.StatusOK, "./.fixtures/organization_response_get_by_id.json"))
+
+	factory := &factory.Factory{
+		HTTPClient: client,
+		Env: &environment.Env{
+			URLAPILedger: URIAPILedger,
+		},
+	}
+
+	org := NewOrganization(factory)
+
+	result, err := org.Get(limit, page)
+	assert.NoError(t, err)
+	assert.NotNil(t, result)
+
+	for i, v := range result.Items {
+		assert.Equal(t, expectedResult.Items[i].ID, v.ID)
+		assert.Equal(t, expectedResult.Items[i].Metadata, v.Metadata)
+	}
+	assert.Equal(t, expectedResult.Limit, limit)
+	assert.Equal(t, expectedResult.Page, page)
+
+	info := httpmock.GetCallCountInfo()
+	assert.Equal(t, 1, info["GET http://127.0.0.1:3000/v1/organizations?limit=5&page=1"])
 }