+++ components/mdz/pkg/cmd/ledger/list_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 	"time"
 
+	"github.com/LerianStudio/midaz/common/mmodel"
 	"github.com/LerianStudio/midaz/components/mdz/internal/domain/repository"
-	"github.com/LerianStudio/midaz/components/mdz/internal/model"
 	"github.com/LerianStudio/midaz/components/mdz/pkg/factory"
 	"github.com/LerianStudio/midaz/components/mdz/pkg/iostreams"
 	"github.com/LerianStudio/midaz/components/mdz/pkg/ptr"
@@ -34,54 +34,50 @@ func Test_newCmdLedgerList(t *testing.T) {
 	cmd := newCmdLedgerList(&ledFactory)
 	cmd.SetArgs([]string{"--organization-id", organizationID})
 
-	list := model.LedgerList{
+	list := mmodel.Ledgers{
 		Page:  1,
 		Limit: 5,
-		Items: []model.LedgerItems{
+		Items: []mmodel.Ledger{
 			{
 				ID:             "0192e362-b270-7158-a647-7a59e4e26a27",
 				Name:           "Ankunding - Paucek",
 				OrganizationID: organizationID,
-				Status: &model.LedgerStatus{
-					Code:        ptr.StringPtr("ACTIVE"),
+				Status: mmodel.Status{
+					Code:        "ACTIVE",
 					Description: ptr.StringPtr("Teste Ledger"),
 				},
 				CreatedAt: time.Date(2024, 10, 31, 16, 22, 29, 232078000, time.UTC),
 				UpdatedAt: time.Date(2024, 10, 31, 16, 22, 29, 232078000, time.UTC),
 				DeletedAt: nil,
-				Metadata: &model.LedgerMetadata{
-					Bitcoin: ptr.StringPtr("3HH89s3LPALardk1jLt2PcjAJng"),
-					Boolean: ptr.BoolPtr(true),
-					Chave:   ptr.StringPtr("metadata_chave"),
-					Double:  ptr.Float64Ptr(10.5),
-					Int:     ptr.IntPtr(1),
+				Metadata: map[string]any{
+					"bitcoin": "3HH89s3LPALardk1jLt2PcjAJng",
+					"chave":   "metadata_chave",
+					"boolean": true,
 				},
 			},
 			{
 				ID:             "0192e258-2c81-7e37-b6ba-a2007495c652",
 				Name:           "Zieme - Mante",
 				OrganizationID: organizationID,
-				Status: &model.LedgerStatus{
-					Code:        ptr.StringPtr("ACTIVE"),
+				Status: mmodel.Status{
+					Code:        "ACTIVE",
 					Description: ptr.StringPtr("Teste Ledger"),
 				},
 				CreatedAt: time.Date(2024, 10, 31, 11, 31, 22, 369928000, time.UTC),
 				UpdatedAt: time.Date(2024, 10, 31, 11, 31, 22, 369928000, time.UTC),
 				DeletedAt: nil,
-				Metadata: &model.LedgerMetadata{
-					Bitcoin: ptr.StringPtr("329aaP47xTc8hQxXB92896U2RBXGEt"),
-					Boolean: ptr.BoolPtr(true),
-					Chave:   ptr.StringPtr("metadata_chave"),
-					Double:  ptr.Float64Ptr(10.5),
-					Int:     ptr.IntPtr(1),
+				Metadata: map[string]any{
+					"bitcoin": "329aaP47xTc8hQxXB92896U2RBXGEt",
+					"chave":   "metadata_chave",
+					"boolean": true,
 				},
 			},
 			{
 				ID:             "0192e257-f5c0-7687-8534-303bae7aa4aa",
 				Name:           "Lang LLC",
 				OrganizationID: organizationID,
-				Status: &model.LedgerStatus{
-					Code:        ptr.StringPtr("ACTIVE"),
+				Status: mmodel.Status{
+					Code:        "ACTIVE",
 					Description: nil,
 				},
 				CreatedAt: time.Date(2024, 10, 31, 11, 31, 8, 352409000, time.UTC),
@@ -92,19 +88,17 @@ func Test_newCmdLedgerList(t *testing.T) {
 				ID:             "0192e251-328d-7390-99f5-5c54980115ed",
 				Name:           "Romaguera and Sons",
 				OrganizationID: organizationID,
-				Status: &model.LedgerStatus{
-					Code:        ptr.StringPtr("ACTIVE"),
+				Status: mmodel.Status{
+					Code:        "ACTIVE",
 					Description: ptr.StringPtr("Teste Ledger"),
 				},
 				CreatedAt: time.Date(2024, 10, 31, 11, 23, 45, 165229000, time.UTC),
 				UpdatedAt: time.Date(2024, 10, 31, 11, 23, 45, 165229000, time.UTC),
 				DeletedAt: nil,
-				Metadata: &model.LedgerMetadata{
-					Bitcoin: ptr.StringPtr("1iR2KqpxRFjLsPUpWmpADMC7piRNsMAAjq"),
-					Boolean: ptr.BoolPtr(false),
-					Chave:   ptr.StringPtr("metadata_chave"),
-					Double:  ptr.Float64Ptr(10.5),
-					Int:     ptr.IntPtr(1),
+				Metadata: map[string]any{
+					"bitcoin": "1iR2KqpxRFjLsPUpWmpADMC7piRNsMAAjq",
+					"chave":   "metadata_chave",
+					"boolean": false,
 				},
 			},
 		},